Pull request for Database branch (#70)
* Update .gitignore

update .gitignore to exclude node_modules

* Add server class to connect to database from web browser

Added a server class to connect to the database from the web browser. Also updated the database class slightly

* Tiny fix

Tiny fix fixing the links in the top menu.

* Search function added

Test search functionality added. Works with searching for conductor for now. Will add more filtering until later.

BREAKING CHANGE
Had to move filter.js from /Code since the relative search paths could not work when debugging locally. Will have to look in to this at a later stage.

* Search functionality for demo

Not a complete solution. But with this implementation it is possible to search for works that are connected to one specific conductor.

* Update filter-page.html

Minor fixes in directories for github server hosting thingy to work

Update filter-page.html

* Add upload matrix

Add upload-matrix

Add upload matrix for easier uploading to the database.

Added id to the search button in filter-page.html

Change port